Add password handling support
Now you can rely on kafo that it creates secure password and store it
into answer file in encrypted form. It's also provided in puppet and
it defines decrypt function in puppet module so you can decrypt these
parameters in your templates like this

<%= scope.function_decrypt($your::password) -%>

In order to mark some parameter as a password you just set it type to
password in documentation.