Progress bar support
We now display progress of installation. When it starts we show progress
bar and percents of progress.

Monkey patch tries to be as safe as possible. We detect Puppet version
and use progress bar only when it's known and supported version. Also
wrapping methods passes all arugments and blocks passed to themselves.

Includes Puppet 3.3 support.