Fix for dumping undef values in puppet 2.6
There are different results for lookupvar of undefined parameter in
puppet 2.6, 2.7 and 3. In puppet 2.6 by default you receive empty
string, in 2.7 you got :undefined symbol and in 3 you get nil.

Luckily you can imitate 2.7 behavior in 2.6 by passing second argument
to lookupvar like this, lookupvar('something', false).
But be careful since in 2.7 and 3 a second argument is expected
to be a hash.