fixes #14450 - replace validation functions with internal impls (#137)
Currently kafo_parsers returns a list of validation functions using the
Puppet AST, and kafo loads these from the modulepath and executes them
in a mock Puppet environment.

This reimplements the common stdlib functions in kafo, removing the need
to set up and run Puppet functions themselves. The code accepts either
the existing AST objects or simpler objects from the parser that respond
to #name and #arguments, allowing this to be switched out in
kafo_parsers later (so the Puppet library isn't needed at all).