fixes #15071 - add --reset-* parameter options (#152)
Parameters can now be reset to the default value set in the manifest by
passing --reset-mod-param or by using a reset sub-menu in interactive
mode.

The behaviour of the Param class has changed to remember defaults.
Previously for a parameter whose default was $::module::params::foo,
the 'default' always remained that parameter name, Puppet was run to
fetch the value of the params class variable, then 'value' was set to
the value of the params variable.

Now 'default' is changed to the value of the params variable so we can
differentiate between the user's value and the default. When resetting
a value to the default it now goes back to the params value instead of
the params variable name.