fixes #14945 - use Hiera to load Puppet classes and parameters (#151)
Replaces use of parser functions to load and transform the answers file
by including classes with hiera_include() and using Hiera with data
binding to pass class parameters in.

This requires Puppet 3 for both hiera_include() and data binding
support.

It allows a custom Hiera config file to be set by `:hiera_config`, and
by adding `kafo_answers` to the hierarchy, parameters can be overridden
or set in a way decided by the installer developer. This will be useful
to set parameters of classes _not_ managed by Kafo, that are indirectly
included.