Fixes #16301 - add key-value store (#171)
The store is loaded from <CONFIG_DIR>/../store.d/. The content is
contained in a yaml files that are loaded in alphabetical order.
The store is read-only during the run-time.
The content can be extended by installing files in the store.d
(e.g. from installer plugins). The instance of the store is in
KafoConfigure.store and can be used form hooks.

Standlaone scripts need to create its own instance by providing the
path. The store class lightweight and is not dependent on Kafo.