fixes #17090 - add acceptance tests for kafo-configure runs
Tests both general behaviour of kafo-configure and kafo-export-params by
generating a new installer project and using the local copy of Kafo. The
tests are not designed to be extensive or to replace unit tests, but to
offer basic reassurance that the installer functionality in
KafoConfigure works.

The speed is poor, mostly due to using Bundler for the installer and
Puppet, which means a development environment with additional
dependencies and fewer performance enhancements such as parser caches
and CFacter, so the number of tests should be kept to a minimum.