fixes #4428 - automatically map namespaced classes to correct paths (#191)
Changes the default locations of namespaced classes (e.g. mymod::foo) to
match the Puppet autoloader locations, assuming the first component is
the module name and classes are laid out in the regular structure. Only
for non-standard layouts would the paths/names need to be overridden,
which modern Puppet versions have reduced support for in any case.