fixes #6158 - only dump defaults for param defaults starting `$`
Refactors how defaults from the manifest and the defaults dump are
stored, keeping them in separate variables and expecting any params
with other variables as the default to have `$` prefixes.

Only params with a `$` default prefix will be dumped, so any other
default string values will not be dumped and are persisted correctly.