fixes #14946 - add --[no-]parser-cache argument to force cache (#200)
With --parser-cache, outdated cache entries will be forced if present,
even if a parser is available. With --no-parser-cache, cache entries
will be ignored even if they appear up to date (same as the existing
kafo-export-params argument).

The default behaviour with neither argument remains the same, to only
use the cache when it appears current.