Fixes #21063 - Refactor fact handling
This moves facts to a BaseContext so it can be used both in migrations
and in hooks.

The facter output is now JSON rather than YAML.

With this the facts are now recursively processed so we can support
structered facts with symbols as well. This might be a breaking change
if users are already using structured facts.