Overhaul loading Hiera answers
The new way of loading answers in Hiera uses a fact to point it to a
temporary file. This may or may not exist, depending on whether
ExecutionEnvironment#store_answers was called.

While this means a breaking change for the consumers using Hiera, it
avoids rewriting the whole Hiera config and makes it much more
predictable.

This also means the Hiera hierachy is always loaded, also during default
determination. The effect is that a Hierachy with defaults can be
introduced which makes the --reset-PARAMETER work reliable to get the
correct defaults when they differ from the module defaults.