Only show scenario options when multiple are available
This removes the option to disable and enable scenarios when there's
only a single one available. That assumes the scenario is already
enabled.

If there is only one, there is also no reason to switch scenarios. That
makes force and compare not needed.

The --scenario and --list-scenarios are kept in because instructions can
refer to an explicit scenario. That keeps the installation manual
consistent.