Fixes #29965 - Support highline 2.x
The biggest change in version 2 is dropping Highline::SystemExtensions
in favor of Highline::Terminal.

Another big change is that the global $terminal is no longer set.
Instead, it has HighLine.default_instance.

It also changes the highline import to only highline instead of
highline/import where possible. This means that not everywhere there is
a global agree/ask/choose/say. Since the scenario manager still uses
highline/import, for hooks there is no change.

It still keeps highline 1.x support because that's shipped on Debian 10
and Ubuntu 18.04. It also allows for a smoother migration since other
gems also use highline.