Determine color usage dynamically
This changes the color usage to be determined dynamically. Prior to this
the config was always read. That meant that if the scenario was ever
written to disk (using save_configuration), it stored the value of
colors_possible? and uses that indefinitely.

After this change, the value of colors will remain nil (determined at
runtime) until the user uses --[no-]colors, which is then persisted.