Add ability to declare modules that can be disabled
Modules in an answer file are assumed to be able to be enabled
and subsequently disabled. However, turning off a puppet module
entirely can produce negative affects for users depending on the
scenario.

This adds the ability to declare specifically the set of modules
that can be disabled. If any are declared (or an empty list) then
the available app options are adjusted. Any disabled module can be
will get a --enable option, any currently enabled option will not
have any option for enable or disable and any module declared in
the disablable modules list will get a --[no-]enable option for
turning it on and off.