Fixes #31304: set exit code before post hooks
If the exit code is not set until the exit handler runs, post hooks
can incorrectly surmise the current exit code. The puppet run sets
the exit code after execution. Users of Kafo such as the foreman-installer
rely on the exit_code in some of the post hooks to determine behavior.