Fixes #31332: Do not display log output when using --help
Logs are buffered before logging configuration is set up, such that
once they are they are added to the log file and to the terminal
output when using verbose mode. That could be triggered when
running --help causing the user to see some log output which
is confusing when using --help. This skips log setup if --help
is present and also prevents new log files from being created
unnecessarily.