Fixes #30365 - Resolve symlinks to determine AIO Puppet
Commit 695ad18f7f3dccb07d93604701f68345c665b181 attempted to deal with
symlinks in /usr/bin or /usr/local/bin, but broke things because paths
in /opt/puppetlabs/bin are also symlinks. That's why it was reverted in
eacdf43c780ef5ac7fab90f0ce3d49b04539bba3.

This uses File.realpath to resolve symlinks so /usr/local/bin/puppet is
fine as long as it is pointing to a file in /opt/puppetlabs.