Fixes #32300 - Do not include hiera configs when loading default values
If hiera configs contain looks for values that do not exist
when loading default values, those looked up values can be empty
string and trigger failures when validating types. Not including
the hiera configs when loading default values should not impact
default value look up since it is designed to get the values from
the puppet modules and not any additional layers.