Add ability to declare classes that can be disabled
This adds configuration of class enablement into the configuration
file along with the ability to specify whether a class can be disabled
or not. This continues to support enable declarations within the answers
file but is superseded by anything declared in the configuration file.

This separates the scenario configuration, from the actual puppet class
parameters found in the answers file. Further, this allows the answers
file to focus on parameters and being part of the standard hiera heirarchy
while declaring properties of the classes within the configuration.

This should be thought of splitting scenario properties, aspects of
a scenario a user should not change, from the elements that should be
configurable via user input.