Fix function that fetches a module by name to use identifier
The previous implementation relied on the name of the module to be
specified which is a specially constructured replacement of colons
with underscores. The hook context calls on this method and indicates
using the identifier as a typicaly Puppet user would encounter it.
For example, my_example::foo would be specified rather than a user
thinking to convert this to my_example_foo. Thus this worked for
single level classes, this would have failed for multilevel
classes.