Ability to specify enabled puppet classes through :classes config
Provide the ability to specify if a class is enabled or disabled through
the classes configuration option rather than through the answer file.
This is a progressive option, as it takes precedence over values in the
answer file but continues to respect answer file if no data is found
in the classes option.