Ensure buffered log messages contain the proper timestamp
Before logging is configured, log messages are stored in an internal
buffer to be dumped at the point at which logging is configured. At
the time the message was being dumped the timestamp was being calculated
making it appear as if a set of log messages all happened at the same
time. This gave no sense that those initial log messages and operations
might be taking long time in some cases.

This moves to storing those buffered messages as LogEvent's that
capture the time of creation such that when those messages are then
dumped to the logs they contain the timestamp of when the message
was actually created rather than the timestamp of when the message
was dumped.