Implement Sensitive data type support
The Sensitive data type is implemented by rewriting Variant to a common
type.

One of the biggest changes is that in order to use Sensitive an entry in
Hiera to convert must exist. This should be implemented in modules using
data in modules but it can be implemented in the installer's Hiera
config as well.

This is based on Puppet's documentation[1].

[1]: https://puppet.com/docs/puppet/6/securing-sensitive-data.html