Add a method MigrationContext.keep_migration for multi-run migrations
Such a migration may alter configuration and answers, or it may noop, but is not required to be recorded in applied migrations and may run more than once, until some condition is met. One simple application would be a backwards compatible migration that noops until a certain target version is reached.