force libdir to be part of our execution environment
When the kafo is executed on a system that itself is a client to a Puppet
server, plugins that got synced (using pluginsync) from said server into
/opt/puppetlabs/puppet/cache/lib/ are picked up instead of the ones in
the configured modules dir under certain circumstances.

For example, puppetlabs-postgresql changed the postgresql_conf provider
from "parsed" to "ruby" in 10.0. If the parsed provider is present in
/opt/puppetlabs/puppet/cache/lib/ it gets used, instead of the new ruby
one, but the calling code (being from puppetlabs-postgresql 10+) doesn't
expect that and fails to properly (yet silently!) configure PostgreSQL.